Fix incorrect value typing when forcing price in SubscriptionTypePaymentItem
Bug introduced in commit 96b4080f

remp/crm#3191

+++ src/Models/PaymentItem/SubscriptionTypePaymentItem.php
@@ -86,7 +86,7 @@ final class SubscriptionTypePaymentItem implements PaymentItemInterface
         return $this;
     }
 
-    public function forcePrice(int $price): self
+    public function forcePrice(float $price): self
     {
         $this->price = $price;
         return $this;
